Move renderMiddleware to server/render
- Rename renderMiddleware -> middleware as it is prefixed by /render
